+++ lib/src/platform/web.dart
@@ -1,14 +1,16 @@
-import 'dart:js_interop';
+import 'dart:async' show Completer, Future;
+import 'dart:js_interop' show JSAny, JSAnyOperatorExtension, JSArray, NumToJSExtension;
+import 'dart:math' show Random;
+import 'dart:typed_data' show ByteBuffer, Uint8List;
 
+import 'package:flutter/material.dart';
+import 'package:n_image_picker/n_image_picker.dart' show ImageController;
 import 'package:web/web.dart' as web;
-import 'dart:math';
-
-import 'package:file_picker/file_picker.dart';
-import 'package:http/http.dart';
-import 'package:flutter/foundation.dart';
-import 'package:n_image_picker/src/platform/helpers.dart';
-import 'package:n_image_picker/src/platform/tools.dart';
-import 'package:n_image_picker/src/response_model.dart';
+import 'package:file_picker/file_picker.dart' show PlatformFile;
+import 'package:http/http.dart' show Response;
+import 'package:n_image_picker/src/platform/helpers.dart' show Helpers;
+import 'package:n_image_picker/src/platform/tools.dart' show PlatformTools;
+import 'package:n_image_picker/src/response_model.dart' show ResponseModel;
 // import 'dart:js' as js;
 
 PlatformTools getInstance() => WebFile();
@@ -74,4 +76,70 @@ class WebFile implements PlatformTools{
     size  : bytes?.length??0,
     bytes : maxSize != null ? bytes != null ? Helpers().Rezize(bytes: bytes, maxSize: maxSize, extension: extension) : null : bytes
   );
+
+  @override
+  void dragAndDrop({required final ImageController controller, required final String className}) async {
+    final body = web.document.body?.getElementsByClassName(className).item(0) as web.HTMLElement;
+
+    // final body = web.window.document.body!;
+    body
+    ..onDragEnter.listen((event) {
+      event.preventDefault();
+      controller.changeOnDragState(true);
+    })
+    ..onDragOver.listen((event) {
+      event.preventDefault();
+      controller.changeOnDragState(false);
+    })
+    ..onDragOver.listen((event) => event.preventDefault() )
+    ..onDrop.listen((event) async {
+      event.preventDefault();
+      web.File? file = (event as web.DragEvent).dataTransfer?.files.item(0);
+      if (file != null){
+        final Completer<Uint8List> completer = Completer<Uint8List>();
+        final web.FileReader reader = web.FileReader();
+
+        reader
+        ..onLoadEnd
+          .listen((event) {
+            if (reader.result != null && reader.result is ByteBuffer) {
+              completer.complete((reader.result as ByteBuffer).asUint8List());
+            }
+          })
+          .onError((event) { null; })
+        ..readAsArrayBuffer(file);
+
+        await completer.future.then((data) =>
+          controller.setFromBytes(
+            name      : file.name,
+            extension : file.name.split(".").last,
+            bytes     : data
+          )
+        );
+      }
+    });
+  }
+
+  @override
+  void createDiv(final GlobalKey widgetKey, {required final String className}) {
+    final div = web.document.createElement('div');
+    // final renderObject = widgetKey.currentContext?.findRenderObject();
+    // final position = renderObject!.localToGlobal(Offset.zero);
+    // final size = renderObject!.size;
+    div.setAttribute("class", className);
+    // div.setAttribute(
+    //   "style",
+    //   "position: absolute; "
+    //   "left: ${position.dx}px; "
+    //   "top: ${position.dy}px; "
+    //   "width: ${size.width}px; "
+    //   "height: ${size.height}px; "
+    //   "background-color: rgba(0, 0, 0, 0.1); "
+    //   "z-index: 1000;"
+    // );
+    web.document.body?.append(div);
+  }
+
+  // @override
+  // dynamic getWebElement({required String className}) => web.document.querySelector('.$className');
 }